pyproject.toml
[project]
name = "payment-system"
version = "0.1.0"
description = ""
authors = [
    {name = "EvgeniyAnistratov"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy[asyncio] (>=2.0.49,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-utils @ git+https://github.com/EvgeniyAnistratov/python_utils.git",
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "dishka (>=1.9.1,<2.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "python-multipart (>=0.0.24,<0.0.25)",
]

[tool.fastapi]
entrypoint = "app.main:app"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
